Update statement regarding 3rd checkout replica
In the previous 'Add nodes' section the number of nodes in the managed-ondemand node group was increased from 2 to 3. Events for pending pod: "1 node(s) didn't match pod affinity rules...that the pod didn't tolerate, 2 node(s) didn't match pod anti-affinity rules."

So maybe should read something like to consider the pod affinity rule as well (I guess this also assumes the reader did change the min/desired to 3 in previous section): 

If we check the running pods we can see that the third checkout pod has been placed in a Pending state since two of the nodes already have a pod deployed and the third node does not have a checkout-redis pod running.